+++ projects/ngx-uc/src/lib/ngx-uc-zoom-view/uc-zoom-view-image-manager.ts
@@ -44,6 +44,11 @@ export class UcZoomViewImageManager extends UcZoomViewManager{
   }
 
   initializeViewer(): void {
+    if (this.isInitialized) {
+      // do not reinitialize
+      return;
+    }
+
     const rootImage = this.image;
     this.outerDiv = this.wrapImage(rootImage);
     this.attachListenersToImage(rootImage);